Jni PoC for errored backtraces
