+++ .gitignore
@@ -62,3 +62,7 @@ backend/logs/
 recorder/.vscode/c_cpp_properties.json
 
 recorder/.vscode/settings.json
+
+*.o
+*.so
+BtPrinter.h
