+++ misc/poc_junk/.gitignore
@@ -1,3 +1,6 @@
 thd_targeted
 thd_untargeted
 bt
+*.o
+*.so
+*.class
