+++ misc/poc_junk/error_backtrace_different_compilation_unit_helper.hh
@@ -1,0 +1,10 @@
+struct Foo {
+    void quux();
+};
+
+int foo();
+
+void print_bt();
+
+int bar();
+
